[ML] Adds a11y tests for categorization and recognizer job wizards
