+++ x-pack/test/accessibility/config.ts
@@ -34,6 +34,7 @@ export default async function ({ readConfigFile }: FtrConfigProviderContext) {
       require.resolve('./apps/ingest_node_pipelines'),
       require.resolve('./apps/index_lifecycle_management'),
       require.resolve('./apps/ml'),
+      require.resolve('./apps/ml_anomaly_detection'),
       require.resolve('./apps/transform'),
       require.resolve('./apps/lens'),
       require.resolve('./apps/upgrade_assistant'),
