+++ x-pack/test/functional/services/ml/index.ts
@@ -36,6 +36,7 @@ import { MachineLearningJobWizardCommonProvider } from './job_wizard_common';
 import { MachineLearningJobWizardCategorizationProvider } from './job_wizard_categorization';
 import { MachineLearningJobWizardMultiMetricProvider } from './job_wizard_multi_metric';
 import { MachineLearningJobWizardPopulationProvider } from './job_wizard_population';
+import { MachineLearningJobWizardRecognizerProvider } from './job_wizard_recognizer';
 import { MachineLearningJobWizardGeoProvider } from './job_wizard_geo';
 import { MachineLearningLensVisualizationsProvider } from './lens_visualizations';
 import { MachineLearningNavigationProvider } from './navigation';
@@ -123,6 +124,7 @@ export function MachineLearningProvider(context: FtrProviderContext) {
     context,
     commonFieldStatsFlyout
   );
+  const jobWizardRecognizer = MachineLearningJobWizardRecognizerProvider(context, commonUI);
   const jobWizardCommon = MachineLearningJobWizardCommonProvider(
     context,
     commonUI,
@@ -200,6 +202,7 @@ export function MachineLearningProvider(context: FtrProviderContext) {
     jobWizardGeo,
     jobWizardMultiMetric,
     jobWizardPopulation,
+    jobWizardRecognizer,
     lensVisualizations,
     mlNodesPanel,
     navigation,
